finish

+++ src/components/Restaurants/Restaurants.jsx
@@ -3,30 +3,44 @@ import React, { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { db } from '../../config/firebase';
 import './restaurants.scss'
-
+import Swal from 'sweetalert2'
 
 const Restaurants = () => {
 
     const [restaurantes, setRestaurantes] = useState([]);
     const restaurantCollectionRef = collection(db, "restaurantes");
-    // const [deletee, setDeletee] = useState("")
 
     const deleteRestaurant = async (id) => {
+        Swal.fire({
+            title: 'Silmək istədiyinizə əminsiniz?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sil',
+            cancelButtonText : 'Ləğv et'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Silindi!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+                const userDoc = doc(db, "restaurantes", id);
+                 deleteDoc(userDoc);
+            }
+        })
 
-        console.log((id));
-        const userDoc = doc(db, "restaurantes", id);
-        let dlt = await deleteDoc(userDoc);
         // setDeletee(dlt)
     };
 
     useEffect(() => {
         const getRestaurantes = async () => {
             const data = await getDocs(restaurantCollectionRef);
-            console.log(data)
             setRestaurantes(data.docs.map((doc) => ({ ...doc.data(), id: doc.id })));
         };
         getRestaurantes();
-    }, []);
+    }, [restaurantes]);
     return (
         <div id='restaurant'>
             <div className='container my-5'>
@@ -36,10 +50,10 @@ const Restaurants = () => {
                         restaurantes.map((restaurant) => {
                             console.log(restaurant)
                             return (
-                                <div className="col-lg-3 my-2">
+                                <div className="col-lg-4 ">
                                     <div className="card">
                                         <div className="card-body">
-                                            <img width='220' src={restaurant.thumbImage} alt="" />
+                                            <img width="100%" src={restaurant.thumbImage} alt="" />
                                             <h5 className="card-title text-center">
                                                 {restaurant.name}
                                             </h5>
